Move EncryptionUils to separate package
